when using useRef for timeline from gasp. it is not updating
when I want the gsap timeline update to a new value, it didn't update a new value of properties when using useRef for gsap timlien. If it was resigned again new gsap tilmeline not the value of properties, it is working. I need to find the reason why it was not working.