I will separate svgFramePack  and put it into its parent
I separate li and its children because I think it works together, but when it hoverover, state didn't update to the componnent. when put dependancy of the state, it works, but all it's children rerender all together. I think svgFrame pack is a just dive, so I will take it out of its children put it into  li tage, parent element.