I'm done with loading gltf file
I tried to use gltfjsx file, but it kept erroring that it is class construct, so I need to put 'new'.

I think it is not class construct, it is functional construct, and I put new, but it didn't work. I think there were some problems on converting gltf to jsx