I remove adding fonts from javascript in Head component and add it in scss file.
when loading fonts, it didn't load fonts  earlier than loading pages, so it was weird.

So I need to load font in scss.
there is the way to use envrionment, but it takes long time to learn and would be heavy.

I just created variable to use it manually for running on local and exporting for the website.